requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.2
frozenlist==1.4.0
idna==3.6
multidict==6.0.4
py-cord==2.4.1
PyYAML==6.0.1
typing_extensions==4.9.0
yarl==1.9.4